@@ -7,14 +7,16 @@
 
 import { Component, Inject, LOCALE_ID, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { faEnvelope, IconDefinition } from '@fortawesome/free-solid-svg-icons';
 import { Person } from '@portal/models/person/person.model';
 import { Search } from '@portal/models/search.model';
 import { SearchService } from '@portal/services/search.service';
 import { SingleItemService } from '@portal/services/single-item.service';
 import { TabChangeService } from '@portal/services/tab-change.service';
 import { UtilityService } from '@shared/services/utility.service';
-import { take } from 'rxjs';
+
+import { Observable, of } from 'rxjs';
+import { delay, map, switchMap, take } from 'rxjs/operators';
+
 import { DOCUMENT } from '@angular/common';
 
 type Field = { key: string; label?: string };
@@ -30,6 +32,8 @@ export class SinglePersonComponent implements OnInit {
   searchTerm: string;
   tabData: any;
   tab = 'person';
+
+  isEmailVisible = false;
 
   affiliationsCaption = $localize`:@@affiliations:Affiliaatiot`;
   educationCaption = $localize`:@@education:Koulutus`;
@@ -84,12 +88,12 @@ export class SinglePersonComponent implements OnInit {
     { key: 'otherNames', label: 'Muut nimet' },
   ];
 
-  person: Person;
+  person$: Observable<Person>;
+  isLoaded$: Observable<boolean>;
 
   initialItemCount = 3;
 
   maxPublicationCount = this.initialItemCount;
-  showAllPublications = false;
 
   maxDatasetCount = this.initialItemCount;
   showAllDatasets = false;
@@ -108,10 +112,25 @@ export class SinglePersonComponent implements OnInit {
     private utilityService: UtilityService,
     private tabChangeService: TabChangeService,
     @Inject(LOCALE_ID) protected localeId: string,
-    @Inject(DOCUMENT) private document: any
+    @Inject(DOCUMENT) private document: any,
   ) {}
 
   ngOnInit(): void {
+    this.person$ = this.route.params.pipe(switchMap((params) => {
+      const id = params["id"];
+
+      return this.singleItemService.getSinglePerson(id).pipe(map((search) => {
+        return search.persons[0] as Person;
+      }));
+    }))
+
+    this.person$.pipe(take(1)).subscribe({
+      complete: () => {
+        // delay masks very fast loading where "404" flashes on screen
+        this.isLoaded$ = of(true).pipe(delay(100));
+      }
+    });
+
     this.route.params.pipe(take(1)).subscribe((params) => {
       this.searchService.searchTerm = params.id;
       this.getData(params.id);
@@ -136,8 +155,7 @@ export class SinglePersonComponent implements OnInit {
         this.responseData = result;
 
         const personRes = result.persons[0];
-        this.person = personRes;
-        this.person.orcidLink = 'https://orcid.org/' + this.person.orcid;
+
         if (personRes) {
           this.setTitle(
             `${personRes.name} - ${$localize`:@@appName:Tiedejatutkimus.fi`}`
@@ -146,9 +164,14 @@ export class SinglePersonComponent implements OnInit {
       });
   }
 
-  showEmail(event, address) {
-    const span = this.document.createElement('span');
-    span.innerHTML = address;
-    event.target.replaceWith(span);
+  showEmail() {
+    this.isEmailVisible = true;
+  }
+
+  showMorePublications() {
+    // Take one value from person and assign its publication length to maxPublicationCount
+    this.person$.pipe(take(1)).subscribe((person) => {
+      this.maxPublicationCount = person.publications.length;
+    });
   }
 }
